building-ai-agents/src/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=1.6.3",
    "requests>=2.32.0",
    "boto3>=1.42.94",
    "strands-agents>=1.37.0",
    "strands-agents-tools>=0.5.1",
]
